DatDF just barely works now
Had a horrible time figuring out how to load an instance of class from __new__ override because the pickle.load also uses the __new__ to initiate so I was running into infinite loops. Ended up resorting to inspect to check if call to __new__ is from within __new__ to change behaviour, but it's pretty messy now. Anyway, it does work at least! So can start working on more things now.