Adding support for data from multiple experiments
Changed Main_Config to be much easier to load up different experiment settings, and to be able to change between them more easily.

Added some rough outlines Experiment specific configs for Mar19 and Sep19 which are unfinished.

Minor chages to core and coreutil

TODO: Need to make it easier to switch configs temporarily for loading single dats for example.