make_dat_standard, and datfactory now work with full datdf instead of dfname
Should still work with old scripts but will print a deprication warning.
In future, full datdf objects should be passed to make_dat_standard so that config issues shouldn't be a problem.
Also full setupdf can be passed along with full config module.
Because only one setupdf, it doesn't make a difference if setupdf is passed at the moment.