Added C.DatHandler, CU.get_alpha() (lever arm), and srs.Constants file
DatHandler is a smarter way to load dats. It keeps track of what dats have been called in a given session and will prevent reloading them if asked for a second time.
Get_alpha just returns the lever arm based on either a single temp and broadening in mV, or an array of data from a theta vs fridge temp.
Constants file is for things like kB etc.