Enormous weekend commit
Should have broken this up way more...
Massive amount of changes, and things are finally starting to be testable with unitttests.
Rewritten a lot of DatHandler and things to do with making dats.
Rethought idea of how to initialize dats. There should be a fairly minimal function which creates the datHDF and puts in the ExpConfig info, and then after that, a DatHDF object should be able to be initialized from an HDF file alone.

Just starting to get this up and running now