Fixed loading Nrg fits from HDF (not perfect)
This might cause a problem if a dNdT nrg fit is saved to HDF because when it is loaded it will assume it is 'I_sense'.
Don't see a particularly easy fix for this other than making a separate function specifically for other NRG funcs. Probably that is the way to go...