Beginning to clean up, refactor, and just make notes to myself about this last CD
A couple of files cleaned up and moved to analysis_tools for use in future CDs.
Mostly just writing notes in the Feb2021 analysis.py files to get a sense of what is useful in there to be saved for future experiments and generally writing notes on what the files were for.